Updated modules/fire.js (added 1 line, deleted 1 line).

+++ modules/fire.js
@@ -22,7 +22,7 @@ var five_year_fires = function(yr) {
   var ymin5 = date.advance(-4, 'year').get('year');
 
   
-  if (max_NBAC_fires_year < y) {
+  if (max_NBAC_fires_year.int() < y) {
     console.log('warning: image year beyond range of fire masks');
   }
